uploading files
There're not assets and mp3s used for this app.

This app was made in group. My part was the player, artist info and map, but indeed I realized the complete integration of all parts, fixes all bugs, create all the data structure to use for all (I tried the easiest way because one of the group doesn't know much about programming) and constants in my views.